Move: 都道府県取得処理を別のファイルに分離した
